refactor: import shared functions from ccbr_tools
