+++ pyproject.toml
@@ -36,6 +36,7 @@ classifiers = [
 requires-python = ">=3.11"
 dependencies = [
     "argparse",
+    "ccbr_tools@git+https://github.com/CCBR/Tools",
     "Click >= 8.1.3",
     "PySimpleGui < 5",
     "snakemake >= 7, < 8",
